feat: always rebuild local source
If `FFI_BUILD_FROM_SOURCE=1` is set, always rebuild from local source.

Change the scripts so that it is built within the `rust` directory
and not on a temporary directory that is different on every rebuild.
This way you get incremental builds if you make changes to the local
source (which is probably your intention if you build from source).