feat: switch to an FFI-level config
This means we don't construct a v3 config at the top level, then try to
convert it to a v2 config. When we do that, we run into issues if, e.g.,
v3 doesn't define a price-list for a version only supported by v2.

This PR also has a bit of cleanup:

1. Removes the manifest CID.
2. Changes the network version from u64 to u32 (simplifies some code).
3. Deletes all the overlay blockstore logic (was only used if the
   manifest was explicitly passed).

Actor "overrides" should now only be done via "redirects".