fix: install aarch64-apple-darwin toolchain on x86 (#421)
We always create universal binaries when compiling on x86. This means
that the aarch64-apple-darwin also needs to be installed. Install that
toolchain automatically if it doesn't exist yet.

This change has also been tested for the case when the toolchain already
exists. In that case it just reports:

    info: component 'rust-std' for target 'aarch64-apple-darwin' is up to date

Thanks @lanzafame for reporting this issue.