fix: remove ClearLayerData call
The `ClearLayerData` FFI call was accidentally introduced with the
Synthetic PoRep. The call does under the hood exactly what `ClearCache`
is doing. It was already removed from Lotus in [PR 11352], so that
we can remove it here as well (and as a next steps from the Rust
side). This reduces the API surface, which is generally a good idea.

[PR 11352]: https://github.com/filecoin-project/lotus/pull/11352